+++ frontend-vue/src/components/bounded-context/namespace/ContextureNamespaces.vue
